Add evaluation functionality
This commit provides the ability for users to evaluate efficiency.

Package.json has been updated to include the `@rails/ujs` dependency for use with AJAX in Rails applications.

An 'Evaluation' controller and associated views have been added to handle the display and calculation of evaluation data.

The routes file has been updated to include routes for the Evaluation controller.

A new file `application.js` has been added in the JavaScript directory for use with `@rails/ujs`.

Tests and YAML files for the Evaluation controller have also been created. The Gemfile.lock has also been edited to add the dotenv-rails gem.

A new method 'effectiveness' has been added in the Evaluation controller to evaluate the effectiveness based on the passed parameters.

Yarn-error.log file is added to track any errors encountered in yarn.