Update evaluation controller and views for better output readability
Changed variables in the evaluation controller to instance variables to provide access in views. The result variables are now rounded for better readability in the UI. UI now provides more detailed results, divided into various levels for further understanding of the evaluation process. Stylesheet for table viewing also added to improve UX. Results are then handled in the js.erb view for async update. This changes will be allowing users to comprehend the output in a simpler way and improves interface interactivity.