Risk assessment calculation implementation (#1)
* Add risk evaluation to app

Modified the existing Rails app by adding routes, controller actions, views and supporting calculations for risk evaluation. The new feature allows the application to perform a complex risk analysis, displaying the data in a visual table format. The purpose of this change is to give users more comprehensive data to help them make informed decisions. The new logical components are highly modular and reusable, promoting the DRY principle for future developers.

* Refactor evaluation_controller & risk_evaluator, update risk view

This commit includes several changes:

1. In `evaluation_controller`, the parameters are now converted to unsafe hashes. This change aims to avoid an error when the form receives hashes instead of plain values in the parameters.

2. In the `risk_evaluator` concern, the loop over keys is replaced by iterating over the hash. This change simplifies the code and makes it more Ruby-like.

3. The calculated risk result is captured more accurately to handle edge cases where the result falls outside the expected range.

4. In the `_risk.html.erb` partial, a label has been changed to more accurately represent the evaluated risk type.

* Add detailed results and enhance code clarity in risk assessment feature

This commit introduces more information to the results of the risk assessment breakdown. Key elements like term assessments, reliability assessments, estimated membership, and membership level are now clearly accounted for and have also been implemented in the generated view. Sections of the assessment process have been renamed for clarity as well. This change was necessary to give users a more in-depth understanding of the risk assessment process. Changes were also made to the styling for better readability and user experience. The corresponding controller and HTML files have been modified to accommodate these changes.

Also, small improvements have been made to the code for enhance its clarity and readability, making the feature easier to maintain in the long run.