Update evaluation data and form input types
The input data for evaluators has been updated in the 'evaluation_controller.rb' and changes have been made to the way inputs are accepted in '_risk.html.erb' and '_effectiveness.html.erb'.
The data change improves the accuracy of evaluations. Changing the input type to a number with a step of 0.1 will allow more precise inputs from the evaluators and labels have been added to the submit buttons for better user understanding.